requirements.txt
Flask==3.1.1
gunicorn==23.0.0
gevent==25.5.1
setuptools==80.9.0

Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg[binary]==3.2.9

celery==5.5.3
redis==6.4.0
flower==2.0.1

alpaca-py==0.42.0
pandas==2.3.1
numpy==2.3.2
TA-Lib==0.6.5

lightgbm==4.6.0
xgboost==3.0.3
scikit-learn==1.7.1
joblib==1.5.1

google-genai==1.31.0
feedparser==6.0.11
requests==2.32.3

prometheus-client==0.22.1
sentry-sdk==2.34.1
python-json-logger==3.3.0
prometheus-flask-exporter==0.23.2

python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

pytest==8.4.1
pytest-flask==1.3.0
pytz==2024.1